requirements.test.txt
-r requirements.txt

pytest-homeassistant-custom-component==0.13.205
pytest-cov>=4.1.0
